+++ controllers/past-grants/views/index.njk
@@ -111,7 +111,7 @@
 
                         <div class="grants-search__filters">
                             {# Javascript component version of filters #}
-                            <grant-filters
+                            <grants-filters
                                 :facets="facets"
                                 :filters="filters"
                                 :status="status"
